Se avanzo con los controladores

+++ src/main/java/FinalProject/Service/ClienteService.java
@@ -7,7 +7,7 @@ import FinalProject.Model.Cliente;
 public interface ClienteService {
 
     public void createCliente(Cliente cliente);
-    public Cliente finByID(Integer id);
+    public Cliente finByID(Long id);
     public List<Cliente> findClientesByNombre(String nombreCliente);
     public List<Cliente> findAllClientes();
